requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
cloudinary==1.33.0
dj-database-url==2.0.0
Django==4.2.2
django-multiselectfield==0.1.12
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.6.3
urllib3==1.26.16
whitenoise==6.5.0
